added login button function
